--- Lesson_6/1.py
@@ -1,36 +1,0 @@
-'''
-Создать класс TrafficLight (светофор) и определить у него один атрибут color (цвет) и метод running (запуск). 
-Атрибут реализовать как приватный. В рамках метода реализовать переключение светофора в режимы: красный, желтый, зеленый. 
-Продолжительность первого состояния (красный) составляет 7 секунд, второго (желтый) — 2 секунды, третьего (зеленый) — на ваше усмотрение. 
-Переключение между режимами должно осуществляться только в указанном порядке (красный, желтый, зеленый). 
-Проверить работу примера, создав экземпляр и вызвав описанный метод.
-
-Задачу можно усложнить, реализовав проверку порядка режимов, 
-и при его нарушении выводить соответствующее сообщение и завершать скрипт.
-'''
-
-class TrafficLight():
-
-    red_pause1 = 7
-    yellow_pause = 2
-    green_pause = 12
-    
-
-    def __init__(self, timer) -> None:
-        self.timer = timer
-
-
-    # def red_pause(self):
-    #     print("red pause = 7")
-
-    # def running():
-
-red = TrafficLight('2222')
-# red.red_pause()
-print(red.timer)
-
-# yellow = TrafficLight('2')
-# print(yellow.yellow_pause)
-
-# green = TrafficLight('12')
-# print(green.green_pause)
